backend/requirements.txt
fastapi
uvicorn
httpx
websockets
pydantic
python-dotenv
bambulabs-api
python-multipart
python-multipart